Update the geocoder parser spec tests
	-split the spec test to match with the split of the parser and decoder
classes
	
		-GeoJsonParserSpec
			-update class names
			-move decoder tests to JsonDecoderSpec class
			-add default constructor var to tests
			
		-JsonDecoderSpec
			-move decoder tests from GeoJsonParserSpec class
			-update class names
